ibd file redesign
