+++ protocol/flows/src/ibd/flow.rs
@@ -1,9 +1,10 @@
 use crate::{
     flow_context::FlowContext,
     flow_trait::Flow,
-    v5::ibd::{HeadersChunkStream, TrustedEntryStream},
+    ibd::{HeadersChunkStream, TrustedEntryStream},
 };
 use futures::future::{join_all, select, try_join_all, Either};
+use itertools::Itertools;
 use kaspa_consensus_core::{
     api::BlockValidationFuture,
     block::Block,
@@ -22,7 +23,8 @@ use kaspa_p2p_lib::{
     dequeue_with_timeout, make_message,
     pb::{
         kaspad_message::Payload, RequestAntipastMessage, RequestHeadersMessage, RequestIbdBlocksBodiesMessage,
-        RequestPruningPointAndItsAnticoneMessage, RequestPruningPointProofMessage, RequestPruningPointUtxoSetMessage,
+        RequestIbdBlocksMessage, RequestPruningPointAndItsAnticoneMessage, RequestPruningPointProofMessage,
+        RequestPruningPointUtxoSetMessage,
     },
     IncomingRoute, Router,
 };
@@ -36,17 +38,18 @@ use tokio::time::sleep;
 use super::{progress::ProgressReporter, HeadersChunk, PruningPointUtxosetChunkStream, IBD_BATCH_SIZE};
 
 /// Flow for managing IBD - Initial Block Download
-pub struct IbdFlowV7 {
+pub struct IbdFlow {
     pub(super) ctx: FlowContext,
     pub(super) router: Arc<Router>,
     pub(super) incoming_route: IncomingRoute,
+    pub(super) body_flow_permitted: bool,
 
     // Receives relay blocks from relay flow which are out of orphan resolution range and hence trigger IBD
     relay_receiver: JobReceiver<Block>,
 }
 
 #[async_trait::async_trait]
-impl Flow for IbdFlowV7 {
+impl Flow for IbdFlow {
     fn router(&self) -> Option<Arc<Router>> {
         Some(self.router.clone())
     }
@@ -69,9 +72,15 @@ struct QueueChunkOutput {
 }
 // TODO: define a peer banning strategy
 
-impl IbdFlowV7 {
-    pub fn new(ctx: FlowContext, router: Arc<Router>, incoming_route: IncomingRoute, relay_receiver: JobReceiver<Block>) -> Self {
-        Self { ctx, router, incoming_route, relay_receiver }
+impl IbdFlow {
+    pub fn new(
+        ctx: FlowContext,
+        router: Arc<Router>,
+        incoming_route: IncomingRoute,
+        relay_receiver: JobReceiver<Block>,
+        body_flow_permitted: bool,
+    ) -> Self {
+        Self { ctx, router, incoming_route, relay_receiver, body_flow_permitted }
     }
 
     async fn start_impl(&mut self) -> Result<(), ProtocolError> {
@@ -180,22 +189,35 @@ impl IbdFlowV7 {
             // means it's in its antichain (because if `highest_known_syncer_chain_hash` was in
             // the pruning point's past the pruning point itself would be
             // `highest_known_syncer_chain_hash`). So it means there's a finality conflict.
-            // TODO: consider performing additional actions on finality conflicts in addition to disconnecting from the peer (e.g., banning, rpc notification)
+            //
+            // TODO (relaxed): consider performing additional actions on finality conflicts in addition
+            // to disconnecting from the peer (e.g., banning, rpc notification)
             return Ok(IbdType::None);
         }
 
         let hst_header = consensus.async_get_header(consensus.async_get_headers_selected_tip().await).await.unwrap();
-        if relay_header.blue_score >= hst_header.blue_score + self.ctx.config.pruning_depth
-            && relay_header.blue_work > hst_header.blue_work
-        {
-            if unix_now() > consensus.async_creation_timestamp().await + self.ctx.config.finality_duration() {
+        // [Crescendo]: use the post crescendo pruning depth depending on hst's DAA score.
+        // Having a shorter depth for this condition for the fork transition period (if hst is shortly before activation)
+        // is negligible since there are other conditions required for activating an headers proof IBD. The important
+        // thing is that we eventually adjust to the longer period.
+        let pruning_depth = self.ctx.config.pruning_depth().get(hst_header.daa_score);
+        if relay_header.blue_score >= hst_header.blue_score + pruning_depth && relay_header.blue_work > hst_header.blue_work {
+            // [Crescendo]: switch to the new *shorter* finality duration only after sufficient time has passed
+            // since activation (measured via the new *larger* finality depth).
+            // Note: these are not critical execution paths so such estimation heuristics are completely ok in this context.
+            let finality_duration_in_milliseconds = self
+                .ctx
+                .config
+                .finality_duration_in_milliseconds()
+                .get(hst_header.daa_score.saturating_sub(self.ctx.config.finality_depth().upper_bound()));
+            if unix_now() > consensus.async_creation_timestamp().await + finality_duration_in_milliseconds {
                 let fp = consensus.async_finality_point().await;
                 let fp_ts = consensus.async_get_header(fp).await?.timestamp;
-                if unix_now() < fp_ts + self.ctx.config.finality_duration() * 3 / 2 {
+                if unix_now() < fp_ts + finality_duration_in_milliseconds * 3 / 2 {
                     // We reject the headers proof if the node has a relatively up-to-date finality point and current
                     // consensus has matured for long enough (and not recently synced). This is mostly a spam-protector
                     // since subsequent checks identify these violations as well
-                    // TODO: consider performing additional actions on finality conflicts in addition to disconnecting from the peer (e.g., banning, rpc notification)
+                    // TODO (relaxed): consider performing additional actions on finality conflicts in addition to disconnecting from the peer (e.g., banning, rpc notification)
                     return Ok(IbdType::None);
                 }
             }
@@ -219,7 +241,7 @@ impl IbdFlowV7 {
 
         let pruning_point = self.sync_and_validate_pruning_proof(&staging_session, relay_block).await?;
         self.sync_headers(&staging_session, syncer_virtual_selected_parent, pruning_point, relay_block).await?;
-        staging_session.async_validate_pruning_points().await?;
+        staging_session.async_validate_pruning_points(syncer_virtual_selected_parent).await?;
         self.validate_staging_timestamps(&self.ctx.consensus().session().await, &staging_session).await?;
         self.sync_pruning_point_utxoset(&staging_session, pruning_point).await?;
         Ok(())
@@ -231,7 +253,11 @@ impl IbdFlowV7 {
         // Pruning proof generation and communication might take several minutes, so we allow a long 10 minute timeout
         let msg = dequeue_with_timeout!(self.incoming_route, Payload::PruningPointProof, Duration::from_secs(600))?;
         let proof: PruningPointProof = msg.try_into()?;
-        debug!("received proof with overall {} headers", proof.iter().map(|l| l.len()).sum::<usize>());
+        info!(
+            "Received headers proof with overall {} headers ({} unique)",
+            proof.iter().map(|l| l.len()).sum::<usize>(),
+            proof.iter().flatten().unique_by(|h| h.hash).count()
+        );
 
         let proof_metadata = PruningProofMetadata::new(relay_block.header.blue_work);
 
@@ -271,7 +297,7 @@ impl IbdFlowV7 {
 
         // Check if past pruning points violate finality of current consensus
         if self.ctx.consensus().session().await.async_are_pruning_points_violating_finality(pruning_points.clone()).await {
-            // TODO: consider performing additional actions on finality conflicts in addition to disconnecting from the peer (e.g., banning, rpc notification)
+            // TODO (relaxed): consider performing additional actions on finality conflicts in addition to disconnecting from the peer (e.g., banning, rpc notification)
             return Err(ProtocolError::Other("pruning points are violating finality"));
         }
 
@@ -302,7 +328,7 @@ impl IbdFlowV7 {
                 .spawn_blocking(move |c| {
                     let ref_proof = proof.clone();
                     c.apply_pruning_proof(proof, &trusted_set)?;
-                    c.import_pruning_points(pruning_points);
+                    c.import_pruning_points(pruning_points)?;
 
                     info!("Building the proof which was just applied (sanity test)");
                     let built_proof = c.get_pruning_point_proof();
@@ -333,13 +359,13 @@ impl IbdFlowV7 {
                 .clone()
                 .spawn_blocking(move |c| {
                     c.apply_pruning_proof(proof, &trusted_set)?;
-                    c.import_pruning_points(pruning_points);
+                    c.import_pruning_points(pruning_points)?;
                     Result::<_, ProtocolError>::Ok(trusted_set)
                 })
                 .await?;
         }
 
-        // TODO: add logs to staging commit process
+        // TODO (relaxed): add logs to staging commit process
 
         info!("Starting to process {} trusted blocks", trusted_set.len());
         let mut last_time = Instant::now();
@@ -352,7 +378,7 @@ impl IbdFlowV7 {
                 last_time = now;
                 last_index = i;
             }
-            // TODO: queue and join in batches
+            // TODO (relaxed): queue and join in batches
             staging.validate_and_insert_trusted_block(tb).virtual_state_task.await?;
         }
         info!("Done processing trusted blocks");
@@ -412,6 +438,14 @@ impl IbdFlowV7 {
             progress_reporter.report_completion(prev_chunk_len);
         }
 
+        if consensus.async_get_block_status(syncer_virtual_selected_parent).await.is_none() {
+            // If the syncer's claimed sink header has still not been received, the peer is misbehaving
+            return Err(ProtocolError::OtherOwned(format!(
+                "did not receive syncer's virtual selected parent {} from peer {} during header download",
+                syncer_virtual_selected_parent, self.router
+            )));
+        }
+
         self.sync_missing_relay_past_headers(consensus, syncer_virtual_selected_parent, relay_block.hash()).await?;
 
         Ok(())
@@ -452,7 +486,7 @@ impl IbdFlowV7 {
         if consensus.async_get_block_status(relay_block_hash).await.is_none() {
             // If the relay block has still not been received, the peer is misbehaving
             Err(ProtocolError::OtherOwned(format!(
-                "did not receive relay block {} from peer {} during block download",
+                "did not receive relay block {} from peer {} during header download",
                 relay_block_hash, self.router
             )))
         } else {
@@ -502,7 +536,7 @@ staging selected tip ({}) is too small or negative. Aborting IBD...",
     }
 
     async fn sync_missing_block_bodies(&mut self, consensus: &ConsensusProxy, high: Hash) -> Result<(), ProtocolError> {
-        // TODO: query consensus in batches
+        // TODO (relaxed): query consensus in batches
         let sleep_task = sleep(Duration::from_secs(2));
         let hashes_task = consensus.async_get_missing_block_body_hashes(high);
         tokio::pin!(sleep_task);
@@ -560,34 +594,55 @@ staging selected tip ({}) is too small or negative. Aborting IBD...",
         let mut jobs = Vec::with_capacity(chunk.len());
         let mut current_daa_score = 0;
         let mut current_timestamp = 0;
-        self.router
-            .enqueue(make_message!(
-                Payload::RequestIbdBlocksBodies,
-                RequestIbdBlocksBodiesMessage { hashes: chunk.iter().map(|h| h.into()).collect() } //change to Request IbdBodyOnlyMessage
-            ))
-            .await?;
-        for &expected_hash in chunk {
-            let msg = dequeue_with_timeout!(self.incoming_route, Payload::IbdBlockBody)?;
-            let blk_header = consensus.async_get_header(expected_hash).await?;
-            let mut blk_body: Vec<Transaction> = vec![];
-            for el in msg.transactions {
-                // a bit roundabout, but into doesn't work nicely
-                blk_body.push(el.try_into()?);
-            }
-            if blk_body.len() == 0 {
-                return Err(ProtocolError::OtherOwned(format!("sent empty block body for block {}", expected_hash)));
-            }
+        if self.body_flow_permitted {
+            self.router
+                .enqueue(make_message!(
+                    Payload::RequestIbdBlocksBodies,
+                    RequestIbdBlocksBodiesMessage { hashes: chunk.iter().map(|h| h.into()).collect() } //change to Request IbdBodyOnlyMessage
+                ))
+                .await?;
+            for &expected_hash in chunk {
+                let msg = dequeue_with_timeout!(self.incoming_route, Payload::IbdBlockBody)?;
+                let blk_header = consensus.async_get_header(expected_hash).await?;
+                let mut blk_body: Vec<Transaction> = vec![];
+                for el in msg.transactions {
+                    // a bit roundabout, but into doesn't work nicely
+                    blk_body.push(el.try_into()?);
+                }
+                if blk_body.is_empty() {
+                    return Err(ProtocolError::OtherOwned(format!("sent empty block body for block {}", expected_hash)));
+                }
 
-            let block = Block { header: blk_header, transactions: blk_body.into() };
-            // if block.hash() != expected_hash {
-            //     return Err(ProtocolError::OtherOwned(format!("expected block {} but got {}", expected_hash, block.hash())));
-            // }
+                let block = Block { header: blk_header, transactions: blk_body.into() };
+                // if block.hash() != expected_hash {
+                //     return Err(ProtocolError::OtherOwned(format!("expected block {} but got {}", expected_hash, block.hash())));
+                // }
 
-            current_daa_score = block.header.daa_score;
-            current_timestamp = block.header.timestamp;
-            jobs.push(consensus.validate_and_insert_body(block).virtual_state_task);
+                current_daa_score = block.header.daa_score;
+                current_timestamp = block.header.timestamp;
+                jobs.push(consensus.validate_and_insert_body(block).virtual_state_task);
+            }
+        } else {
+            self.router
+                .enqueue(make_message!(
+                    Payload::RequestIbdBlocks,
+                    RequestIbdBlocksMessage { hashes: chunk.iter().map(|h| h.into()).collect() }
+                ))
+                .await?;
+            for &expected_hash in chunk {
+                let msg = dequeue_with_timeout!(self.incoming_route, Payload::IbdBlock)?;
+                let block: Block = msg.try_into()?;
+                if block.hash() != expected_hash {
+                    return Err(ProtocolError::OtherOwned(format!("expected block {} but got {}", expected_hash, block.hash())));
+                }
+                if block.is_header_only() {
+                    return Err(ProtocolError::OtherOwned(format!("sent header of {} where expected block with body", block.hash())));
+                }
+                current_daa_score = block.header.daa_score;
+                current_timestamp = block.header.timestamp;
+                jobs.push(consensus.validate_and_insert_block(block).virtual_state_task);
+            }
         }
-
         Ok(QueueChunkOutput { jobs, daa_score: current_daa_score, timestamp: current_timestamp })
     }
 }
