+++ protocol/flows/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod flow_context;
 pub mod flow_trait;
 pub mod flowcontext;
+pub mod ibd;
 pub mod service;
 pub mod v5;
 pub mod v6;
