+++ protocol/flows/src/v5/request_pruning_point_and_anticone.rs
@@ -13,7 +13,7 @@ use kaspa_p2p_lib::{
 };
 use log::debug;
 
-use crate::{flow_context::FlowContext, flow_trait::Flow, v5::ibd::IBD_BATCH_SIZE};
+use crate::{flow_context::FlowContext, flow_trait::Flow, ibd::IBD_BATCH_SIZE};
 
 pub struct PruningPointAndItsAnticoneRequestsFlow {
     ctx: FlowContext,
