+++ protocol/flows/src/v5/request_pruning_point_utxo_set.rs
@@ -1,4 +1,4 @@
-use crate::{flow_context::FlowContext, flow_trait::Flow, v5::ibd::IBD_BATCH_SIZE};
+use crate::{flow_context::FlowContext, flow_trait::Flow, ibd::IBD_BATCH_SIZE};
 use itertools::Itertools;
 use kaspa_consensus_core::errors::consensus::ConsensusError;
 use kaspa_core::debug;
