+++ protocol/flows/src/v6/request_pruning_point_and_anticone.rs
@@ -16,7 +16,7 @@ use kaspa_p2p_lib::{
 use log::debug;
 use std::sync::Arc;
 
-use crate::{flow_context::FlowContext, flow_trait::Flow, v5::ibd::IBD_BATCH_SIZE};
+use crate::{flow_context::FlowContext, flow_trait::Flow, ibd::IBD_BATCH_SIZE};
 
 pub struct PruningPointAndItsAnticoneRequestsFlow {
     ctx: FlowContext,
