--- protocol/flows/src/v7/ibd/mod.rs
@@ -1,7 +1,0 @@
-mod flow;
-mod negotiate;
-mod progress;
-mod streams;
-
-pub use flow::*;
-pub use streams::*;
